rename filter item to sort item
